integrated Region objects so now we can store all of the data we need into them.
I need to get these made when we run the floodfill,  that way we can store all our data as we go.

I made a GetRegionFromTile() func, but it's obvs slow having to iterate over tiles.
I made a Region field in the Tile class, so we can set what region a tile is in when we run the region floodfill:   so we would create the Region, then set that starting tile's Region to that new Region, and add each successive tile.
After the floodfill, those Regions get passed off to the ConvexHull code, which right now gives back coords of points, but it will be setting those coords into a field on each of those Region objects.