Add IRoleNeeded and PlayerInfo
Partial implementation of a new RoleNeeded interface for the tankSteer and Turret script. The PlayerInfo is a data holder right now with a nasty singleton while we're still doing debug stuff.
I added some extra debug controls and started stubbing out the Turret class, along with update for the tankSteer so it uses the new interface and doesn't accept input when PlayerInfo doesn't have the role picked.