Start Projectile work
I made a prefab for the shell with the needed scripts, placed one in the scene so I could start messing with it, and started putting together some classes that I can use to start pew pewing on the map soon. My brain fried right now, so I'll return to this tomorrow and see what all I can do to get this branch ready to go.