Underlying array change. New version of (#6)
* Map spawn entry points are being found.

Looping around the border and getting the midpoints of spans of flat ground.

* REEEEE

* reeeeeee cant find pathfinding

* Got Astar pathfinding working.

Still need to add land cost w/ noisemap, but it's finding paths.

* oops had a for loop messed up

* Pathfinding working, backtracking not implemented

Need to implement backtracking so it doesn't get stuck.

* Ok we're pathfinding!!! Woohoo!!!

I  still need to have it show the shortest path, but we stay winning.

* tried to integrate badPaths, its not doing what I want

reeee

* idk fam

* yayyyy! (no progress lmao)

* Pathfinding is implemented and working.

I still need to figure out the no-path exit.

* need to fix the FindLandmassFloodFill func

It should color all the low squares blue if it works. It's running till it runs out of memory :/

* yeetus

cleetus

* Start of LandmassMarking

I tried to floodfill all the regions I found, but instead I got a pretty checkboard pattern on the colormap.
I think the problem lies in the Floodfill function, but it's possible it could be stemming from the GetNeighbors function.

* added 64x64 option to Terrain Size dropdown

Added 128 and 64 grid options. Also set noise scale correctly so we get the same map.

* commit before I break things

UwU

* No more checkerboard :)

woot woot!

* cleanup

cleaned up print statements (only printing region size if > 1), made terrain 256x256 again, runs in less than a second now that floodfill is fixed.

* Implemented Convex Hull and fixed floodfill issues (mostly).

Sickkkkkk

* cleanup, final before more pathfinding work

added editor script and a few QOL fixes

* Ignoring the minimap texture

Updated the .gitignore so it doesn't try to track the changes to the minimap's texture each time the map is generated

* Update Floodfill debug view

Just renamed the variable that handles showing the marked regions. Also added a tooltip so it's easier to notice it. UwU

* fixed where values were being added to on subsequent runs

we gucci

* Debug messages for neighbors

Trying to track down why we're only getting regions with small sizes, so adding some print statements for now so I can start manually tracking things down until I attach the debugger later.

* Found bug in Astar

Need to have the badPaths list correctly utilized. Need to add to it ONLY when the neighbor in question is too tall. This needs to be done in the updated neighbor function with the elevation helper function.

* Different A* implementation.

It seems that the heuristic of fCost is pretty shit for what I need this for, probs gonna go back to hCost (dist to end).

* Add TileUnderElevation

Adds an inverse for TileOverElevation in case you want to find flat tiles

* renamed func to AddNeighbors.

* Pathfinding.cs cleanup unused funcs

* finish cleanup before starting hull trigonometry calcs

* I FIXED THE REGION FILL!!!

Just make it simpler, that's all I had to do. The floodfill for the regions now works without any issue that I've found in my testing. Getting ready to test the heck out of it now

* let's goooo

I figured out the math for the bisection of the terrain for roads  :D

* uhhh progress of some sort

You ever have one of those "what in the actual fuck are you doing" moments? yeah this is that lolol

* Started the process of generating Arterial main roads.

Basically implemented FindNeighbors(), but for an entire region lol

Need to finish the Regions class, so that we can store info about each region as we create them. That way we can store all the Tiles, bounding boxes, hullpoints, etc. there. @BarndoBoi plz halp with that

* .

* aight that's all I can do for now

fuggin classes and shit REEEEEE

* alright I'm GOING to BED

UwU

* integrated Region objects so now we can store all of the data we need into them.

I need to get these made when we run the floodfill,  that way we can store all our data as we go.

I made a GetRegionFromTile() func, but it's obvs slow having to iterate over tiles.
I made a Region field in the Tile class, so we can set what region a tile is in when we run the region floodfill:   so we would create the Region, then set that starting tile's Region to that new Region, and add each successive tile.
After the floodfill, those Regions get passed off to the ConvexHull code, which right now gives back coords of points, but it will be setting those coords into a field on each of those Region objects.

* got floodfill creating regions.

* Everything should be switched over to the Regions system now.

sick tiddies

* Should have all region data being stored now. Working on the midline calcs.

God I wish I remembered algebra and slopes and shit rn. I gotta go brush up lol

* Finished midpoint calcs. Not all working, but we're getting there.

* alright it's getting the data but it's all wonky

I'll  debug tomorrow :p

* Revert "Merge branch 'master' into roadgen2"

This reverts commit 019d26ccbc2e4107af51619d937aa7a3f5421862, reversing
changes made to e1ea0bdb2c1fbc185d7132b72091feadd2395168.

* yeet

les gooo

* Quicksave

hmmmmm

* ok found the bug. Saving before fucking with things...

SetHeights reads the noiseMap array from a diff axis or some shit?? so we need to make x,y --> y,x in the array we pass it I guess.

* made FetchFloadValue_ReversedYXarray() for Map. Will pass to SetHeights in LayerTerrain.

SetHeights needs an array indexed YX not XY like we were doing. Need to find out what else needs the reversed data, and need to make sure tiles are getting the correct data on them.

* commented out changes to "reset" to how it was.

I'm so glad I figured it outtttttt

* Heights and colors are in the correct places now.

Changed SetHeights to the reversed array that it needs, and changed

colorMap[x * width + y] = color;    --->    colorMap[y * width + x] = color;

in hindisight that colormap was backwards, I should've seen that.

* cleanup layerTerrain and get ready to work on midpoints now that everything is fixed

---------

Co-authored-by: Victoria Eastep <114838742+BarndoBoi@users.noreply.github.com>