IT FUCKING WORKS
I got a build going on my local network and did a lot of reading in the NetworkRoomManager. I had to enable auto player creation, which means I need to make a proper prefab to move the data over to, and the PlayerInfo needs to be adjusted. Should be able to do some good work tomorrow though!