Mostly Bug Free Lobby
I think I've squashed most of the bugs I was encountering with the game setup. The ready button works, the playerinfo is now correct, and everything replicates except for the client's inputs. I'll probably need to give the client authority to fix that, but it's basically ready for main now