Move Enemy Folder and make Types
The enemy folder doesn't really need to be in the network folder, even though several of them will be networked later on.
I also wrote up an enemy so that enemies can have a type and added a field to the enemy class to store the data. Also the radar info got linked up to the enemy :3