Stub NavigateRoads and update Nodes to keep a reference to their renderer
I'm starting work on making the enemies actually follow the roads. I had to go back and mess with the Node class itself to find an easy spot to get the renderer, and now it'll just be a matter of adapting some of the stuff from Sebastian's PathFollower example script. Pausing for a few minutes now though...