Fix navigation rotation bug and spawn order bug
I fixed it so that there are no more errors when spawning in the test enemies now. They were also inheriting a 90 degree z rotation from their path somehow, so I went ahead and cancelled that out so only x,y rotation is taken. The NevigateRoads script is still not fully done though...