Rename DriveRoads and update
I changed the name back to NavigateRoads since it makes more sense and am going to try orientating the unit towards the next node to see if that prevents the wrong turns I saw last time in testing. I also changed some of the squad ScriptableObjects so that they are easier to test. There's also a mock Steering class that is for later