Almost Attempt Server Message Broadcast
Alright, I've been trying to get everything set up so that I can send connection messages. The issue is that I never really set things up for setting the name in the lobby properly, so now I'm dealing with issues there. I'm done for the night now, but this is getting close