Create TeamsScriptableObject and use it
So while we don't have any code for it right now, we have three teams we can use. Enemy Squads now take a Teams asset to use for themselves as well. Just need to test that the teams can assign themselves to the static list without issue in the editor.