Add WinState to project
This is just stubbed out so we have a way to actually check when the game is finished. My idea is that different gamemodes will have different scripts that monitor the game state and interact with this static class. For now this is completely unused and is just so I remember to come back and do it. Also, I made it so that multiple teams can have a victory.