Add TeamScriptableObject list
I changed things up so that the TeamPicker holds which Teams can be joined for now. I need to add more data to this, but I can at least keep testing after this change. I modified the Teams class to use a method for adding teams to make it easier to track my code.